feat: add count-up, reorganize component paths, remove unused stats
