+++ src/components/Home/HomeHero/HomeHeroStats.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Box, Stack, Typography } from "@mui/material";
 import { useTranslation } from "next-i18next";
 import { HomeHeroStatsProps } from "../../../types/Home";
+import CountUp from "react-countup"
 
 const HomeHeroStats = ({ stats = { personalities: 0, claims: 0, reviews: 0 } }: HomeHeroStatsProps) => {
     const { t, i18n } = useTranslation();
@@ -46,7 +47,7 @@ const HomeHeroStats = ({ stats = { personalities: 0, claims: 0, reviews: 0 } }: 
                         component="span"
                         className="home-header-stats-value"
                     >
-                        {item.value.toLocaleString(i18n.language)}
+                        <CountUp start={0} end={item.value} duration={2} separator="." />
                     </Typography>
                     <Typography
                         component="span"
